# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Alicerce API
  version: "1"
  description: >
    Written first (ADR 0011): every response in spec/requests/api/v1 is
    validated against this document by committee, and
    frontend/src/api/schema.d.ts is generated from it.
servers:
  - url: /api/v1
paths:
...
  /categories:
    get:
      operationId: listCategories
      summary: List the current organization's product categories
      description: Every role can read (ADR 0008); owner, admin and purchasing can also write.
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        "200":
          $ref: "#/components/responses/CategoryListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
    post:
      operationId: createCategory
      summary: Create a product category
      description: Owner, admin and purchasing only (ADR 0008).
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          $ref: "#/components/responses/CategoryResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /categories/{id}:
    get:
      operationId: showCategory
      summary: Show a product category
      description: Every role can read (ADR 0008).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateCategory
      summary: Update a product category
      description: Owner, admin and purchasing only (ADR 0008).
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /products:
    get:
      operationId: listProducts
      summary: List the current organization's products
      description: Every role can read (ADR 0008); owner, admin and purchasing can also write.
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: category_id
          in: query
          schema: { type: integer }
        - name: active
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          description: Case-insensitive search over name and sku.
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/ProductListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
    post:
      operationId: createProduct
      summary: Create a product
      description: Owner, admin and purchasing only (ADR 0008).
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          $ref: "#/components/responses/ProductResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /products/{id}:
    get:
      operationId: showProduct
      summary: Show a product
      description: Every role can read (ADR 0008).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          $ref: "#/components/responses/ProductResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateProduct
      summary: Update a product
      description: Owner, admin and purchasing only (ADR 0008).
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          $ref: "#/components/responses/ProductResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
...
  /partners:
    get:
      operationId: listPartners
      summary: List the current organization's partners
      description: Every role can read (ADR 0008); owner, admin and purchasing can also write.
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: customer
          in: query
          schema: { type: boolean }
        - name: supplier
          in: query
          schema: { type: boolean }
        - name: active
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          description: Case-insensitive search over name.
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/PartnerListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
    post:
      operationId: createPartner
      summary: Create a partner
      description: Owner, admin and purchasing only (ADR 0008).
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartnerRequest"
      responses:
        "201":
          $ref: "#/components/responses/PartnerResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /partners/{id}:
    get:
      operationId: showPartner
      summary: Show a partner
      description: Every role can read (ADR 0008).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          $ref: "#/components/responses/PartnerResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updatePartner
      summary: Update a partner
      description: Owner, admin and purchasing only (ADR 0008).
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartnerRequest"
      responses:
        "200":
          $ref: "#/components/responses/PartnerResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: The csrf_token from the most recent GET or POST /session response.
      schema:
        type: string
  responses:
...
    Error:
      description: An error in the shared envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseBody"
...
    CategoryResponse:
      description: A product category.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CategoryResponseBody"
    CategoryListResponse:
      description: A page of the current organization's product categories.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CategoryListResponseBody"
    ProductResponse:
      description: A product.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProductResponseBody"
    ProductListResponse:
      description: A page of the current organization's products.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProductListResponseBody"
...
    PartnerResponse:
      description: A partner.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PartnerResponseBody"
    PartnerListResponse:
      description: A page of the current organization's partners.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PartnerListResponseBody"
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [ code, message, details, request_id ]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
          nullable: true
    ErrorResponseBody:
      type: object
      additionalProperties: false
      required: [ error ]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Meta:
      type: object
      additionalProperties: false
      required: [ page, per_page, total ]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
...
    Unit:
      type: object
      additionalProperties: false
      required: [ id, code, name, active ]
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        active: { type: boolean }
...
    Category:
      type: object
      additionalProperties: false
      required: [ id, name, active ]
      properties:
        id: { type: integer }
        name: { type: string }
        active: { type: boolean }
    CategoryResponseBody:
      type: object
      additionalProperties: false
      required: [ data ]
      properties:
        data:
          $ref: "#/components/schemas/Category"
    CategoryListResponseBody:
      type: object
      additionalProperties: false
      required: [ data, meta ]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Category" }
        meta:
          $ref: "#/components/schemas/Meta"
    CreateCategoryRequest:
      type: object
      additionalProperties: false
      required: [ name ]
      properties:
        name: { type: string }
    UpdateCategoryRequest:
      type: object
      additionalProperties: false
      required: [ name, active ]
      properties:
        name: { type: string }
        active: { type: boolean }
    UnitConversion:
      type: object
      additionalProperties: false
      required: [ purchase_unit, factor ]
      properties:
        purchase_unit:
          $ref: "#/components/schemas/Unit"
        factor:
          type: string
          pattern: '^[0-9]{1,9}\.[0-9]{6}$'
          description: Decimal string, always 6 places (ADR 0006). Stock quantity = purchase quantity x factor.
    Product:
      type: object
      additionalProperties: false
      required: [ id, sku, name, active, category, stock_unit, unit_conversion ]
      properties:
        id: { type: integer }
        sku: { type: string }
        name: { type: string }
        active: { type: boolean }
        category:
          nullable: true
          allOf: [ { $ref: "#/components/schemas/Category" } ]
        stock_unit:
          $ref: "#/components/schemas/Unit"
        unit_conversion:
          $ref: "#/components/schemas/UnitConversion"
    ProductResponseBody:
      type: object
      additionalProperties: false
      required: [ data ]
      properties:
        data:
          $ref: "#/components/schemas/Product"
    ProductListResponseBody:
      type: object
      additionalProperties: false
      required: [ data, meta ]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        meta:
          $ref: "#/components/schemas/Meta"
    CreateProductRequest:
      type: object
      additionalProperties: false
      required: [ sku, name, stock_unit_id, purchase_unit_id, factor ]
      properties:
        sku: { type: string }
        name: { type: string }
        category_id:
          type: integer
          nullable: true
        stock_unit_id: { type: integer }
        purchase_unit_id: { type: integer }
        factor:
          type: string
          description: Decimal string, greater than zero, below 1000000000, at most 6 places (ADR 0006). More places or a larger value is a validation_failed (too_many_decimals, less_than), never a rounding.
    UpdateProductRequest:
      type: object
      additionalProperties: false
      required: [ sku, name, stock_unit_id, purchase_unit_id, factor, active ]
      properties:
        sku: { type: string }
        name: { type: string }
        category_id:
          type: integer
          nullable: true
        stock_unit_id: { type: integer }
        purchase_unit_id: { type: integer }
        factor:
          type: string
          description: Decimal string, greater than zero, below 1000000000, at most 6 places (ADR 0006). More places or a larger value is a validation_failed (too_many_decimals, less_than), never a rounding.
        active: { type: boolean }
...
    Partner:
      type: object
      additionalProperties: false
      required: [ id, name, document_type, document_number, customer, supplier, email, phone, active ]
      properties:
        id: { type: integer }
        name: { type: string }
        document_type:
          type: string
          enum: [ cpf, cnpj ]
        document_number:
          type: string
          description: >
            Digits only for a CPF; digits and upper-case letters for a CNPJ
            (ADR 0012, alphanumeric since July 2026). No punctuation.
        customer: { type: boolean }
        supplier: { type: boolean }
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        active: { type: boolean }
    PartnerResponseBody:
      type: object
      additionalProperties: false
      required: [ data ]
      properties:
        data:
          $ref: "#/components/schemas/Partner"
    PartnerListResponseBody:
      type: object
      additionalProperties: false
      required: [ data, meta ]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Partner" }
        meta:
          $ref: "#/components/schemas/Meta"
    CreatePartnerRequest:
      type: object
      additionalProperties: false
      required: [ name, document_type, document_number, customer, supplier ]
      properties:
        name: { type: string }
        document_type:
          type: string
          enum: [ cpf, cnpj ]
        document_number: { type: string }
        customer: { type: boolean }
        supplier: { type: boolean }
        email: { type: string }
        phone: { type: string }
    UpdatePartnerRequest:
      type: object
      additionalProperties: false
      required: [ name, document_type, document_number, customer, supplier, active ]
      properties:
        name: { type: string }
        document_type:
          type: string
          enum: [ cpf, cnpj ]
        document_number: { type: string }
        customer: { type: boolean }
        supplier: { type: boolean }
        email: { type: string }
        phone: { type: string }
        active: { type: boolean }